Use custom image for linux-static-arm64
Use custom image for static arm64 builds, which is now essentially a
copy of our Dockerfile for dynamically linked arm64 musl builds but with
irrelevant parts removed.

See https://github.com/prisma/team-orm/issues/608 for rationale and
detailed description.

This Dockerfile was tested locally with commands identical to those
`engineer` runs:

- `docker run -it --rm -v $(pwd):/work -w /work -e SQLITE_MAX_VARIABLE_NUMBER=250000 -e SQLITE_MAX_EXPR_DEPTH=10000 -e LIBZ_SYS_STATIC=1 docker.io/prismagraphql/build:linux-static-arm64 bash`
  to run the image
- `cargo build --release -p query-engine --target aarch64-unknown-linux-musl --features vendored-openssl`
  etc to build a specific package

(the only difference is that technically engineer runs it as a single
command without a shell in between and builds each binary in a separate
container but it doesn't matter)

Closes: https://github.com/prisma/team-orm/issues/608