fix: use Amazon Linux base image for glibc builds
CentOS 7 is EOL so the glibc builds are now failed. We can consider
raising the minimum supported glibc version from 2.17 to something more
modern in Prisma 6 but for now we need to maintain compatibility with
the same range of platforms.

This PR changes the base image to Amazon Linux 1, which is the only RHEL
7 derivative which is still maintained.